Bug resolving:
adding the 'global' keyword to simple_tracker_node.py code to solve the problem of not updating the latest position of funnywing.
changing the flags of msgpack functions to 'raw=False' to solve the runtime error in the tests.
Solving the problem of locking in guided mode by simple_tracker_node.py. tested in simulation but not in field test.
Adding the Stabilize mode to the combo box of GUI.