Redesign the Structor of the package:
python packages and modules are moved to the wing_navigator ROS package root, consequently change in setup.py.

Adding a new python package to the wing_navigator to support our new RfCommunication design. Implementing initial sketches of the Jobs and JobFactory. Not Completed yet.

Adding a new node RfConnectionTest.py to test the new Communication design.

TODO: correct the usage and imports related to the wing_modules across the project.