The attitude_to_sensor_msgs_Imu have been tested using simulation. The results of the tests show that, despite the fact that using quaternion_from_euler function from ROS tf module, we get different quaternions than ones from the /mavros/imu/data topic in the RPI side, if we use euler_from_quaternion function(from the ROS tf module) with the same euler sequence rzyx in the GCS side, the transmitted euler angles(which is in ATTITUDE message) will be correct. So regardless of different quaternions we can use euler angles in GUI and so on. Also, the rate of the transmission was good and there was no significant delay.
TODO: In the case of using the attitude in guidance and control, I think we must have same quaternions in both sides of the communication. I have searched alot to make this to happen and found some things about the difference in coordinate frames used in FCU and ROS(ENU and NED) and have tried many euler sequences and off the shelf functions for conversions, but cant get it right.

So TODO is: In future try to get both the same quaternion and euler angles in both communication sides(RPI and GCS) and even you can post questions about this in relevant places.