Solved the bug of having no connection between back and front end. I think front end did not receive the backend signals, since we were setting the backEnd object as the root context property but the signals owned by the class containing the Signals. As I understood, Only the Signals and Slots from the object specified as root context property can be transferred between back and front end. So if we set backEnd as root context property and use signals defined in another class and that other class is a field of the backEnd class, signals would not be transmitted.
To solve this we have created a communication class holding all the signals and slots necessary for communicating between back and front end. Then, created an object of that in the main function and set that object as root context property of the QML engine. Now, any class or code that want to transfer data from back to front ends or vice versa can use this object as input. So the bug solved!