Changing the map qml code to make the markers smaller when zooming in and bigger when zooming out.
TODO:

marker color overlays becomes corrupted when zooming in too much and finally disappears. I think we can solve the color problem with changing the color of marker in photoshop. For disappearing I don't know now.

a problem in qml is that when I define a function I can't return a local values like size determined based on some conditions to the marker size. Solve this later.