Adding Ardupilot parameter configuration controls to GUI.
Note that for inferring the parameter type, the parameter is pulled from the autopilot using mavros, its type is inferred and then the parameter set request is sent to autopilot using mavros.

In front-end base we have used the component for the first time which is a good reference for front-end development.