Bug Solving: solving pymavlink bug occurring on sending mavlink messages that have string fields.
The problem solved by applying the encode method to a list of mavlink messages that have string fields, like PARAM_SET etc. This is done in the mavrosToMavlink method in RfCommunicationHandler class and no change is applied to the RfConnection class which is intended to be used as an external library(mentioned this for software engineering purposes).