Implementation of different Frame captures.
In this commit the frame capture interface is implemented using different libraries:

1. OpenCV

2. OpenCV with Gstreamer backend

3. FFMPEG

4. Gst module

also some minor changes are applied. For example, the frame size is added to the CameraFrameCaptureInterface class and its default values are specified in the implementor classes.

Note: As I tested the runcam6 HDMI output is 4K regardless of what we set it, so many corruptions in my stream display was because of this huge output while my link was capable of transferring 1080P30. Actually, tests using univision camera showed this to us.