Adding the ability to lock on the last detected object.
Using the ID of the last displayed track in the GUI for automation of the lock process. This last track is detected in the frame processor and has been updated by Qt signal mechanism. Then the updated last track id in the back-end will be sent over RF when I click on the Auto Lock button in the GUI. Using the last displayed ID has two benefits. First, there is no need to change codes of sensor block and second, user will see the track on which will be locked(otherwise the last track in the jetson will be locked on and user does not know what exactly it is and this is dangerous). for more details, go through the source code.