Adding launch files for simulation and field test.
added 4 launch files, 2 for simulation and 2 for field test. the field test launch files are incomplete and needs to check the files on the jetson. do it in near future.

using parse_known_args instead of parse_args, to prevent errors when running the system with launch file. the errors caused by the fact that ROS adds some arguments to the list of arguments used for launching the nodes, for example it adds __name argument. so, since we use argparse.ArgumentParser, and these added args are not added in the code to the parser of this module, these auto added args cause the error. using this new method, allows the parser to ignore these unknown arguments.