Resolving warning by filtering out the MAVLink2 messages coming from autopilot.
On jetson, the autopilot uses MAVLink2 by default on telemetry ports(you can configure it to use MAVLink1 and resolve this warning) and so some messages like MCU_STATUS(with ID 11039) causing some warnings, since we use MAVLink1 in the codes and MAVLink1 messages have 8-bit IDs. So by filtering-out messages with ID greater than 255, we have resolved this warning.